+++ ramp-database/rampdb/tools/submission.py
@@ -19,7 +19,6 @@ from ..model import SubmissionFile
 from ..model import SubmissionFileTypeExtension
 from ..model import SubmissionOnCVFold
 
-from .event import is_admin
 from .event import get_event
 from .user import get_user_by_name
 
@@ -779,46 +778,3 @@ def submit_starting_kits(session, event_name, team_name, path_submission):
     # revert the minimum duration between two submissions
     event.min_duration_between_submissions = min_duration_between_submissions
     session.commit()
-
-
-def is_accessible_code(session, event_name, user_name,
-                       submission_name='sandbox'):
-    """Whether or not the user can look at the code submission.
-
-    Parameters
-    ----------
-    session : :class:`sqlalchemy.orm.Session`
-        The session to directly perform the operation on the database.
-    event_name : str
-        The event name.
-    user_name : str
-        The user name.
-    submission_name : str, default == 'sandbox
-        The submission name which you should be shown. Default is the sandbox
-        submission.
-
-    Returns
-    -------
-    is_accessible : bool
-        Whether or not the submission can be shown.
-    """
-    # local import to avoid circular import issue
-    from .team import is_user_signed_up
-    user = get_user_by_name(session, user_name)
-    if not user.is_authenticated or not user.is_active:
-        return False
-    if is_admin(session, event_name, user_name):
-        return True
-    if not is_user_signed_up(session, event_name, user_name):
-        return False
-
-    event = get_event(session, event_name)
-    submission_name = (event.ramp_sandbox_name
-                       if submission_name == 'sandbox' else submission_name)
-    submission = get_submission_by_name(session, event_name, user_name,
-                                        submission_name)
-    if user == submission.event_team.team.admin:
-        return True
-    if event.is_public:
-        return True
-    return False
