+++ ramp-database/rampdb/tools/tests/test_event.py
@@ -39,9 +39,6 @@ from rampdb.tools.event import get_event_admin
 from rampdb.tools.event import get_problem
 from rampdb.tools.event import get_workflow
 
-from rampdb.tools.event import is_accessible_event
-from rampdb.tools.event import is_admin
-
 HERE = os.path.dirname(__file__)
 
 
@@ -236,42 +233,3 @@ def test_check_event(session_scope_function, config):
     delete_event(session_scope_function, 'iris_test')
     event = get_event(session_scope_function, None)
     assert len(event) == 1
-
-
-def test_check_admin(session_toy_db):
-    event_name = 'iris_test'
-    user_name = 'test_iris_admin'
-    assert is_admin(session_toy_db, event_name, user_name)
-    user_name = 'test_user'
-    assert not is_admin(session_toy_db, event_name, user_name)
-    add_event_admin(session_toy_db, event_name, user_name)
-    assert is_admin(session_toy_db, event_name, user_name)
-    event_admin = get_event_admin(session_toy_db, event_name, user_name)
-    assert event_admin.event.name == event_name
-    assert event_admin.admin.name == user_name
-    user_name = 'test_user_2'
-    assert get_event_admin(session_toy_db, event_name, user_name) is None
-
-
-@pytest.mark.parametrize(
-    "event_name, user_name, is_accessible",
-    [('xxx', 'test_iris_admin', False),
-     ('iris_test', 'test_user', False),
-     ('iris_test', 'test_iris_admin', True),
-     ('iris_test', 'test_user_2', True),
-     ('boston_housing_test', 'test_user_2', False)]
-)
-def test_is_accessible_event(session_toy_db, event_name, user_name,
-                             is_accessible):
-    # force one of the user to not be approved
-    if user_name == 'test_user':
-        user = get_user_by_name(session_toy_db, user_name)
-        user.access_level = 'asked'
-        session_toy_db.commit()
-    # force an event to be private
-    if event_name == 'boston_housing_test':
-        event = get_event(session_toy_db, event_name)
-        event.is_public = False
-        session_toy_db.commit()
-    assert is_accessible_event(session_toy_db, event_name,
-                               user_name) is is_accessible
